requirements.txt
numpy
pandas
matplotlib
scikit-learn
joblib
pyserial
